trying to split the long file in 2
