+++ ci/rspec-split-config.json
@@ -1,3 +1,5 @@
 [
-    { "files": ["spec/models/census_employee_spec.rb"]}
+    { "files": ["spec/models/census_employee_spec.rb"]},
+    { "files": ["spec/models/census_employee_part2_spec.rb"]}
+
 ]
